requirements.txt
python-dotenv
openai-agents
sendgrid
autogen-core
autogen-ext==0.6.4
autogen-ext[openai, diskcache, redis, ollama, mcp]
tiktoken
autogen-agentchat
langchain
langchain_community
azure-identity
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-openai
playwright
uv
mcp-server-fetch